+++ dbt/tests/adapter/basic/test_incremental.py
@@ -1,7 +1,7 @@
 import pytest
 
 # TODO: update to new dbt-artifacts once available
-from dbt.artifacts.results import RunStatus
+from dbt.artifacts.schemas.results import RunStatus
 from dbt.tests.adapter.basic import files
 from dbt.tests.util import run_dbt, check_relations_equal, relation_from_name
